Move FrozenOrThawed to its own module
